requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
async-timeout==5.0.1
attrs==25.4.0
azure-cognitiveservices-speech==1.47.0
azure-core==1.37.0
beautifulsoup4==4.14.3
billiard==4.2.4
celery==5.6.0
certifi==2025.11.12
cffi==2.0.0
channels==4.3.2
channels-redis==4.3.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.3
curlify==3.0.0
dataclasses-json==0.6.7
distro==1.9.0
Django==5.2.9
django-admin-interface==0.32.0
django-colorfield==0.14.0
django-cors-headers==4.9.0
django-filter==25.2
django-redis==6.0.0
django-unfold==0.74.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
elevenlabs==2.27.0
environ==1.0
exceptiongroup==1.3.1
facebook_business==24.0.1
faiss-cpu==1.13.1
fastapi==0.124.4
frozenlist==1.8.0
google==3.0.0
google-auth==2.47.0
google-genai==1.57.0
greenlet==3.3.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.6.1
langchain==1.2.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.2.1
langchain-openai==1.1.3
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.0
langsmith==0.4.59
marshmallow==3.26.1
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.3.5
openai==2.12.0
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==24.6.1
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
PyYAML==6.0.3
redis==7.1.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
SQLAlchemy==2.0.45
sqlparse==0.5.4
starlette==0.50.0
tenacity==9.1.2
text-unidecode==1.3
tiktoken==0.12.0
tqdm==4.67.1
twilio==9.9.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
uuid_utils==0.12.0
uvicorn==0.38.0
vine==5.1.0
wcwidth==0.2.14
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
